renamed phase-allocation route to phase
